Fixed positions and sizes with new scaling system
Fixed the story, the HUD, the character customisation screens, and the
death screen. Stopped the game from trying to re-scale if the resolution
is 1280x720, or if the game is full screen. Fixed unrelated problem with
weapon slot. Fixed infinite loop problem with teleportation caused by
new scale system. Also added a test for a way of reloading the script to
test a new change without having to restart the game. Doesn't really
work.