Redone options menu
Added a separate menu items class for option editors. A radio button,
switch, or dropdown menu can be added.Made the fullscreen option a
switch, the resolution option a dropdown menu, and added the option to
switch between basic and smooth scaling. When fullscreen is exited, the
resolution now returns to what it was before. The options are now saved
by clicking on the 'apply' button. Finally, tried to add arrows to the
RPGSystem font, but it didn't really work.